+++ edx_proctoring/settings/common.py
@@ -17,6 +17,7 @@ def plugin_settings(settings):
             'proctoring/js/models/proctored_exam_allowance_model.js',
             'proctoring/js/models/proctored_exam_attempt_model.js',
             'proctoring/js/models/proctored_exam_model.js',
+            'proctoring/js/models/learner_onboarding_model.js',
             'proctoring/js/collections/proctored_exam_allowance_collection.js',
             'proctoring/js/collections/proctored_exam_attempt_collection.js',
             'proctoring/js/collections/proctored_exam_collection.js',
@@ -25,6 +26,7 @@ def plugin_settings(settings):
             'proctoring/js/views/proctored_exam_allowance_view.js',
             'proctoring/js/views/proctored_exam_attempt_view.js',
             'proctoring/js/views/proctored_exam_view.js',
+            'proctoring/js/views/proctored_exam_info.js',
             'proctoring/js/views/proctored_exam_instructor_launch.js',
             'proctoring/js/proctored_app.js',
             'proctoring/js/exam_action_handler.js'